Reorganize package layout to match Android best practices
Move files into a layered + feature-grouped structure:
- LokiLogger → util/
- MusicPlaybackService, MediaActionReceiver, PositionInterpolator → playback/
- KotifyMappers → data/

Pure moves — no code changes, no class renames, no behavior change.
Manifest updated with new fully-qualified class names for the service
and receiver.

Closes #161
