+++ src/app/src/main/java/ch/snepilatch/app/playback/MediaActionReceiver.kt
@@ -1,4 +1,4 @@
-package ch.snepilatch.app
+package ch.snepilatch.app.playback
 
 import android.content.BroadcastReceiver
 import android.content.Context
